requirements.txt
requests==2.20.1
environs
python-telegram-bot==13.11